Move routes file into routes folder
